% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_clusters}
\alias{check_clusters}
\title{Validate Independent-Clusters Assumption}
\usage{
check_clusters(data, cid, id)
}
\arguments{
\item{data}{A data frame.}

\item{cid}{Character. Cluster ID column name.}

\item{id}{Character. Individual ID column name.}
}
\description{
Checks that no individual ID appears in more than one cluster.
Stops with an informative error if the assumption is violated.
}
\keyword{internal}
