% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.patp}
\alias{summary.patp}
\title{Summary Method for patp Objects}
\usage{
\method{summary}{patp}(object, ...)
}
\arguments{
\item{object}{A \code{patp} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A list with named components for further inspection.
}
\description{
Returns the full curve(s) and (if applicable) the test result.
}
